Add Result handling

+++ src/main/kotlin/edu/unh/cs/ai/omab/algorithms/ExpectationMaximization.kt
@@ -6,7 +6,7 @@ import edu.unh.cs.ai.omab.domain.Simulator
 import java.util.stream.IntStream
 
 
-fun expectationMaximization(mdp: MDP, horizon: Int, world: Simulator, simulator: Simulator): Long {
+fun expectationMaximization(mdp: MDP, horizon: Int, world: Simulator, simulator: Simulator): Double {
     var currentState = mdp.startState
 
     return IntStream.iterate(0, { i -> i + 1 }).limit(horizon.toLong()).mapToLong {
@@ -19,5 +19,5 @@ fun expectationMaximization(mdp: MDP, horizon: Int, world: Simulator, simulator:
         currentState = nextState
 
         reward.toLong()
-    }.sum()
+    }.sum().toDouble()
 }
